分离出ImageSynthesisCore
