+++ src/main/java/xmmt/dituon/share/ImageSynthesis.java
@@ -1,16 +1,11 @@
 package xmmt.dituon.share;
 
 
-import javax.imageio.ImageIO;
 import java.awt.*;
-import java.awt.geom.Ellipse2D;
 import java.awt.image.BufferedImage;
-import java.io.IOException;
-import java.net.HttpURLConnection;
-import java.net.URL;
 import java.util.ArrayList;
 
-public class ImageSynthesis {
+public class ImageSynthesis extends ImageSynthesisCore {
 
     public static BufferedImage synthesisImage(BufferedImage sticker,
                                                ArrayList<AvatarModel> avatarList, ArrayList<TextModel> textList,
@@ -68,44 +63,11 @@ public class ImageSynthesis {
         switch (avatar.getPosType()) {
             case ZOOM:
                 g2dDrawZoomAvatar(g2d, avatar.getImage(),
-                        avatar.nextPos(), avatar.getNextAngle(), avatar.isRound());
+                        avatar.nextPos(), avatar.getNextAngle(), avatar.isAntialias());
                 break;
             case DEFORM:
-                g2dDrawDeformAvatar(g2d, avatar.getImage(), avatar.getDeformData());
+                g2dDrawDeformAvatar(g2d, avatar.getImage(), avatar.getDeformData().getDeformPos(), avatar.getDeformData().getAnchor());
         }
-    }
-
-    private static void g2dDrawZoomAvatar(Graphics2D g2d, BufferedImage avatarImage, int[] pos,
-                                          float angle, boolean isRound) {
-        if (avatarImage == null) {
-            return;
-        }
-
-        int x = pos[0];
-        int y = pos[1];
-        int w = pos[2];
-        int h = pos[3];
-        if (angle == 0) {
-            g2d.drawImage(avatarImage, x, y, w, h, null);
-            return;
-        }
-
-        if (isRound || angle % 90 == 0) {
-            BufferedImage newAvatarImage = new BufferedImage(avatarImage.getWidth(), avatarImage.getHeight(), avatarImage.getType());
-            Graphics2D rotateG2d = newAvatarImage.createGraphics();
-            rotateG2d.rotate(Math.toRadians(angle), avatarImage.getWidth() / 2, avatarImage.getHeight() / 2);
-            rotateG2d.drawImage(avatarImage, null, 0, 0);
-            rotateG2d.dispose();
-            g2d.drawImage(newAvatarImage, x, y, w, h, null);
-            return;
-        }
-
-        g2d.drawImage(rotateImage(avatarImage, angle), x, y, w, h, null);
-    }
-
-    private static void g2dDrawDeformAvatar(Graphics2D g2d, BufferedImage avatarImage, AvatarModel.DeformData deformData) {
-        BufferedImage result = ImageDeformer.computeImage(avatarImage, deformData.getDeformPos());
-        g2d.drawImage(result, deformData.getAnchor()[0], deformData.getAnchor()[1], null);
     }
 
     private static void g2dDrawTexts(Graphics2D g2d, ArrayList<TextModel> texts) {
@@ -120,58 +82,4 @@ public class ImageSynthesis {
         }
     }
 
-    public static BufferedImage convertCircular(BufferedImage input, boolean antialias) throws IOException {
-        BufferedImage output = new BufferedImage(input.getWidth(), input.getHeight(), BufferedImage.TYPE_4BYTE_ABGR);
-        Ellipse2D.Double shape = new Ellipse2D.Double(0, 0, input.getWidth(), input.getHeight());
-        Graphics2D g2 = output.createGraphics();
-        g2.setClip(shape);
-
-        if (antialias) {
-            g2.setRenderingHint(RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON);
-        }
-
-        g2.drawImage(input, 0, 0, null);
-        g2.dispose();
-        return output;
-    }
-
-    public static BufferedImage rotateImage(BufferedImage avatarImage, float angle) {
-        double sin = Math.abs(Math.sin(Math.toRadians(angle))),
-                cos = Math.abs(Math.cos(Math.toRadians(angle)));
-        int w = avatarImage.getWidth();
-        int h = avatarImage.getHeight();
-        int neww = (int) Math.floor(w * cos + h * sin),
-                newh = (int) Math.floor(h * cos + w * sin);
-        BufferedImage rotated = new BufferedImage(neww, newh, avatarImage.getType());
-        Graphics2D g2d = rotated.createGraphics();
-        rotated = g2d.getDeviceConfiguration().createCompatibleImage(
-                rotated.getWidth(), rotated.getHeight(), Transparency.TRANSLUCENT);
-        g2d.dispose();
-        g2d = rotated.createGraphics();
-
-        g2d.translate((neww - w) / 2, (newh - h) / 2);
-        g2d.rotate(Math.toRadians(angle), w / 2, h / 2);
-        g2d.drawRenderedImage(avatarImage, null);
-        g2d.dispose();
-        return rotated;
-    }
-
-    public static BufferedImage getAvatarImage(String URL) {
-        HttpURLConnection conn = null;
-        BufferedImage image = null;
-        try {
-            java.net.URL url = new URL(URL);
-            conn = (HttpURLConnection) url.openConnection();
-            conn.connect();
-            image = ImageIO.read(conn.getInputStream());
-            conn.disconnect();
-        } catch (Exception e) {
-            System.out.println("获取头像失败\nHttpURLConnection: " + conn + "\nURL: " + URL);
-            e.printStackTrace();
-        } finally {
-            assert conn != null;
-            conn.disconnect();
-        }
-        return image;
-    }
 }
