+++ src/test/java/xmmt/dituon/example/AbstractTest.java
@@ -1,6 +1,5 @@
 package xmmt.dituon.example;
 
-import kotlin.Pair;
 import org.junit.BeforeClass;
 import xmmt.dituon.share.AvatarExtraDataProvider;
 import xmmt.dituon.share.BasePetService;
@@ -11,11 +10,8 @@ import java.awt.image.BufferedImage;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.InputStream;
-import java.util.List;
 
-public class AbstractTest {
-
-    static BasePetService petService = new BasePetService();
+public abstract class AbstractTest {
     static final int DEFAULT_BUFFER_SIZE = 8192;
     static final String INPUT_ROOT = "./example-data/input/";
     static final String OUTPUT_ROOT = "./example-data/output/";
@@ -34,17 +30,7 @@ public class AbstractTest {
         }
     }
 
-    protected void testGeneral(String saveName, String key, List<TextData> additionTextDatas) {
-        Pair<InputStream, String> resultStreamAndType = petService.generateImage(key,
-                new AvatarExtraDataProvider(() -> avatarImage1, () -> avatarImage2,
-                        null, null)
-                , null, additionTextDatas);
-        String finalSaveName = OUTPUT_ROOT + getClass().getSimpleName() + "-" + saveName + "." + resultStreamAndType.getSecond();
-        copyInputStreamToFile(resultStreamAndType.getFirst(), new File(finalSaveName));
-        System.out.println("test " + key + " done.");
-    }
-
-    private void copyInputStreamToFile(InputStream inputStream, File file) {
+    protected void copyInputStreamToFile(InputStream inputStream, File file) {
         // append = false
         try (FileOutputStream outputStream = new FileOutputStream(file, false)) {
             int read;
